CREATE : 카카오 로그인구현
CREATE : 카카오 로그인 구현(+test.py)

CREATE : 카카오 로그인 구현

CREATE : 카카오 로그인구현

CRAETE : 카카오 로그인 구현

CREATE : 카카오 로그인 구현

CREATE : 카카오 로그인 구현

CREATE : 카카오로그인 구현

UPDATE : 카카오 로그인 수정

WIP

.

수정중

UPDATE : 카카오로그인 리뷰 반영

UPDATE: 카카오로그인 수정

UPDATE: 카카오로그인 에러코드 추가

UPDATE : 카카오 로그인 수정

리뷰 반영

UPDATE : unit test 생성

유닛 테스트 생성

.

+++ users/models.py
@@ -5,7 +5,7 @@ from core.models import TimeStampModel
 # Create your models here.
 class User(TimeStampModel):   
     email    = models.CharField(max_length = 45,unique = True)
-    kakao_id = models.IntegerField(unique = True, null=True)
+    kakao_id = models.BigIntegerField(unique = True, null=True)
     class Meta:
         db_table = 'users'
